+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development] - Unreleased
 
+### Changed
+
+- Raise exceptions instead of returning `None` for our parser
+- Deactivate no yet implemented module (fleet composition) by default
+
 
 ## [0.0.1-alpha.2] - 2023-08-10